+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Manually add audio latency for better playback
 * Add a restart input
 * Add an option to make a savestate each time the thread set has changed
+* Lock inputs in the input editor
 
 ### Changed
 ### Fixed